seven/requirements.txt
argcomplete==1.10.0
astroid==2.4.2
beautifulsoup4==4.8.0
blis==0.4.1
boto==2.49.0
boto3==1.14.3
botocore==1.17.3
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cymem==2.0.3
Cython==0.29.14
datefinder==0.7.1
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
en-core-web-sm
extract-msg==0.23.1
Flask==1.1.2
Flask-Cors==3.0.8
fuzzywuzzy==0.18.0
gensim==3.8.3
idna==2.9
IMAPClient==2.1.0
importlib-metadata==1.6.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.2
nltk==3.5
numpy==1.18.5
olefile==0.46
pandas==1.0.4
pdfminer.six==20181108
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
pycryptodome==3.9.7
pylint==2.5.3
PyMuPDF==1.17.0
pytesseract==0.3.4
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2020.1
regex==2020.6.8
requests==2.23.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.12.0
sklearn==0.0
smart-open==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.2.4
SpeechRecognition==3.8.1
srsly==1.0.2
textract==1.6.3
thinc==7.4.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.46.1
typed-ast==1.4.1
tzlocal==1.5.1
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.9
zipp==3.1.0
